fix: remove code path for running query in worker
