+++ packages/common/src/types/schedulerTaskList.ts
@@ -4,7 +4,6 @@ import { type SchedulerIndexCatalogJobPayload } from './catalog';
 import { type UploadMetricGsheetPayload } from './gdrive';
 import { type RenameResourcesPayload } from './rename';
 import {
-    type AsyncWarehouseQueryPayload,
     type CompileProjectPayload,
     type DownloadCsvPayload,
     type EmailNotificationPayload,
@@ -46,7 +45,6 @@ export const SCHEDULER_TASKS = {
     GENERATE_DAILY_JOBS: 'generateDailyJobs',
     EXPORT_CSV_DASHBOARD: 'exportCsvDashboard',
     RENAME_RESOURCES: 'renameResources',
-    RUN_ASYNC_WAREHOUSE_QUERY: 'runAsyncWarehouseQuery',
     ...EE_SCHEDULER_TASKS,
 } as const;
 
@@ -74,7 +72,6 @@ export interface TaskPayloadMap {
     [SCHEDULER_TASKS.EXPORT_CSV_DASHBOARD]: ExportCsvDashboardPayload;
     [SCHEDULER_TASKS.SLACK_AI_PROMPT]: SlackPromptJobPayload;
     [SCHEDULER_TASKS.RENAME_RESOURCES]: RenameResourcesPayload;
-    [SCHEDULER_TASKS.RUN_ASYNC_WAREHOUSE_QUERY]: AsyncWarehouseQueryPayload;
 }
 
 export interface EETaskPayloadMap {
